\name{dfmeta-package}
\alias{dfmeta-package}
\alias{dfmeta}
\docType{package}

\title{
Meta-Analysis of Phase I Dose-Finding Early Phases Clinical Trials.
}

\description{
Meta-analysis approaches for Phase I dose finding early phases clinical trials in order to better suit requirements in terms of maximum tolerated dose (MTD) and maximal dose regimem (MDR). This package has currently three different approaches: (a) an approach proposed by Zohar et al, 2011 (denoted as ZKO), (b) the Variance Weighted pooling analysis (called VarWT) and, (c) the Random 
Effects Model Based (REMB) algorithm, where user can put his/her own model based approach or use the existing random effect logistic regression model (named as glimem) through 'dfmeta' package.
}

\author{
Artemis Toumazi <artemis.toumazi@gmail.com>, Sarah Zohar <sarah.zohar@inserm.fr>, Anand N. Vidyashankar <avidyash@gmu.edu>, Jie Xu <jxu13@gmu.edu> and Moreno Ursino <moreno.ursino@inserm.fr>

Maintainer: Artemis Toumazi <artemis.toumazi@gmail.com>
}

\references{
Zohar, S. , Katsahian, S. and O'Quigley, J. (2011), An approach to meta analysis of dose finding studies. Statist. Med., <doi:10.1002/sim.4121>.
}

\keyword{package}

