% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{methods}
\name{plot,ZKO,missing-method}
\alias{plot,ZKO,missing-method}
\title{The graphical representation of dose-toxicity curve.}
\usage{
\S4method{plot}{ZKO,missing}(x, y = NA, ...)
}
\arguments{
\item{x}{a "ZKO" object.}

\item{y}{the "y" argument is not used in the plot-method for "ZKO" object.}

\item{\dots}{other arguments to the \code{\link[=graphics]{plot.default}} function can be passed here.}
}
\description{
A plotting function representing graphically the dose-toxicity relationship including the target probability of toxicity. In addition you can see as circles the proportional sizes of the number of patients evaluated for toxicity at each dose level.
}
\references{
Zohar, S. , Katsahian, S. and O'Quigley, J. (2011), An approach to meta analysis of dose finding studies. Statist. Med., <doi:10.1002/sim.4121>.
}
\author{
Artemis Toumazi <artemis.toumazi@gmail.com>, Sarah Zohar <sarah.zohar@inserm.fr>, Anaund N. Vidyashankar <avidyash@gmu.edu>, Jie Xu <jxu13@gmu.edu> and Moreno Ursino <moreno.ursino@inserm.fr>
}
