\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ZKO-method}
\alias{show,REMB-method}

\title{S4 Methods for Function \code{show}}

\description{
S4 Methods for function \code{show}.
}

\section{Methods}{
  \describe{
    \item{\code{signature(object = "ZKO")}}{
    S4 method to store and present the ZKO algorithm's results. 
    }
    
    \item{\code{signature(object = "REMB")}}{
    S4 method to store and present the REMB algorithm's results.
    }
  }
}

\author{
Artemis Toumazi <artemis.toumazi@gmail.com>, Sarah Zohar <sarah.zohar@inserm.fr>, Anand N. Vidyashankar <avidyash@gmu.edu>, Jie Xu <jxu13@gmu.edu> and Moreno Ursino <moreno.ursino@inserm.fr>
}

\keyword{methods}
