% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_classes.R
\docType{class}
\name{ZKO-class}
\alias{ZKO-class}
\title{An S4 class to represent the results using the ZKO algorithm.}
\description{
An S4 class to represent the results using the ZKO algorithm.
}
\section{Slots}{

\describe{
\item{\code{dataTox}}{A data frame containing information for each patient at each simulation and trial.}

\item{\code{target}}{The target of probability of toxicity.}

\item{\code{doses}}{A vector with the doses panel.}

\item{\code{ti}}{A vector including the total number of toxicities at each dose level.}

\item{\code{ni}}{A vector consisting the total number of patients at each dose level.}

\item{\code{realite}}{A vector with the observed toxicity probabilities at each dose level.}

\item{\code{relfreq}}{A vector with the mean relative frequencies of allocation at each dose level.}

\item{\code{estimAlpha}}{The estimated value for variable alpha.}

\item{\code{estimProb}}{The estimated probability of toxicity.}
}}

